Remove connection parameter from call to authorize
The connection parameter is optional but can be used to specify the
connection that should be automatically used when a user logs in using
OAuth. By removing the parameter the user must specify the desired
connection but this will prevent calls from failing if the specified
connection is changed or renamed later.